Type portability: convert from int to double before returning
git-svn-id: file:///home/svn/repos/kite/interpreter/branches/kite-llvm@607 6f0f3af2-bb18-0410-b4af-aa22a44f8c3b

+++ main.cpp
@@ -19,7 +19,7 @@ int main (int argc, char * const argv[]) {
 	InitializeNativeTarget();
 	llvm_start_multithreaded();
 	
-	ConstantValue<double> *constant = new ConstantValue<double>(42.0);
+	ConstantValue<int> *constant = new ConstantValue<int>(42);
 	MethodValue *method = new MethodValue("life_universe_everything");
 	method->push_instruction(constant);
 	CompilerState *state = new CompilerState();
